+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log #
 
+## Unreleased ##
+
+### Fixed ###
+
+  - Quote shell arguments.
+
 ## 0.0.2 - 2015-04-11 ##
 
   - Put the package metadata in the `.el` file.